S.Res. 4 Introduced in Senate
A resolution expressing the sense of the Senate that the Supreme Court of the United States erroneously decided Kennedy v. Louisiana, No. 07-343 (2008), and that the eighth amendment to the Constitution of the United States allows the imposition of the death penalty for the rape of a child.

Version:  Introduced in Senate (is)
Date:     2009-01-06
Congress: 111

Source: https://www.govinfo.gov/content/pkg/BILLS-111sres4is/xml/BILLS-111sres4is.xml

Sponsored-By: V000127