H.Con.Res. 184 Introduced in House
Expressing the sense of the Congress that the bankruptcy proceedings of Lehman Brothers Holding Inc. and Lehman Brothers Europe Inc. be resolved in an equitable and expeditious process between the United States and United Kingdom, and that the interests of American investors be given due consideration and be treated with urgency throughout.

Version:  Introduced in House (ih)
Date:     2009-09-14
Congress: 111

Source: https://www.govinfo.gov/content/pkg/BILLS-111hconres184ih/xml/BILLS-111hconres184ih.xml

Sponsored-By: M001137