H.Res. 976 Reported in House
Providing for consideration of the Senate amendment to the bill (H.R. 3326) making appropriations for the Department of Defense for the fiscal year ending September 30, 2010, and for other purposes; for consideration of the joint resolution (H.J.Res. 64) making further continuing appropriations for fiscal year 2010, and for other purposes; for consideration of the bill (H.R. 4314) to permit continued financing of Government operations; for consideration of the Senate amendment to the bill (H.R. 2847) making appropriations for the Departments of Commerce and Justice, and Science, and Related Agencies for the fiscal year ending September 30, 2010, and for other purposes; and for other purposes.

Version:  Reported in House (rh)
Date:     2009-12-16
Congress: 111

Source: https://www.govinfo.gov/content/pkg/BILLS-111hres976rh/xml/BILLS-111hres976rh.xml

Sponsored-By: P000597
Roll-Call: House 111-1-982 2009-12-16 Passed 235-193
Roll-Call: House 111-1-983 2009-12-16 Passed 228-201