H.Res. 1413 Introduced in House
Expressing the sense of the House of Representatives that the holding in Miranda v. Arizona may be interpreted to provide for the admissibility of a terrorist suspect's responses in an interrogation without administration of the Miranda warnings, to the extent that the interrogation is carried out to acquire information concerning other threats to public safety.

Version:  Introduced in House (ih)
Date:     2010-05-27
Congress: 111

Source: https://www.govinfo.gov/content/pkg/BILLS-111hres1413ih/xml/BILLS-111hres1413ih.xml

Sponsored-By: T000260